flex&bison: Add option noyywrap to wc and calc examples
When a lex scanner reaches the end of yyin (by default STDIN) it calls
yywrap so that it can adjust the value of yyin in case more files are
present. Once all files are exhausted yywrap returns 1. This mechanism
is quite old and is nowadays unpopular. The option noyywrap notifies lex
to not call yywrap.

Signed-off-by: Kristiyan Stoimenov <kristoimenov@gmail.com>